[czp/data] support data loading from different frameworks
