+++ model/torch_model.py
@@ -212,139 +212,6 @@ class Hook():
 # =============================================================================
 
 
-
-# =============================================================================
-# Data processing
-# =============================================================================
-def get_data_npz(data_folder = './', dataset = 'fmnist.npz', val_split = 1/5):
-    '''
-    Args:
-        data_folder : Location of dataset
-        dataset (string): <dataset name>.npz, must have 4 keys -- xtr, ytr, xte, yte
-            xtr: (num_trainval_samples, num_features...)
-            ytr: (num_trainval_samples,)
-            xte: (num_test_samples, num_features...)
-            yte: (num_test_samples,)
-        val_split (float, optional): What fraction of training data to use for validation
-            If not 0, val data is taken from end of training set. Eg: For val_split=1/6, last 10k images out of 60k for MNIST are taken as val
-            If 0, train set is complete train set (including val). Test data is returned as val set
-            Defaults to 1/5
-
-    Returns:
-        xtr (torch tensor): Shape: (num_train_samples, num_features...)
-        ytr (torch tensor): Shape: (num_train_samples,)
-        xva (torch tensor): Shape: (num_val_samples, num_features...). This is xte if val_split = 0
-        yva (torch tensor): Shape: (num_val_samples,). This is yte if val_split = 0
-        xte (torch tensor): Shape: (num_test_samples, num_features...)
-        yte (torch tensor): Shape: (num_test_samples,)
-
-    '''
-    loaded = np.load(data_folder+dataset)
-    xtr = loaded['xtr']
-    ytr = loaded['ytr']
-    xte = loaded['xte']
-    yte = loaded['yte']
-    
-    ## Val split ##
-    if val_split != 0:
-        split = int((1-val_split)*len(xtr))
-        xva = xtr[split:]
-        yva = ytr[split:]
-        xtr = xtr[:split]
-        ytr = ytr[:split]
-    
-    ## Convert to tensors on device ##
-    xtr = torch.as_tensor(xtr, dtype=torch.float, device=device)
-    ytr = torch.as_tensor(ytr, dtype=torch.long, device=device)
-    xva = torch.as_tensor(xva, dtype=torch.float, device=device)
-    yva = torch.as_tensor(yva, dtype=torch.long, device=device)
-    xte = torch.as_tensor(xte, dtype=torch.float, device=device)
-    yte = torch.as_tensor(yte, dtype=torch.long, device=device)
-
-    if val_split != 0:
-        return xtr,ytr, xva,yva, xte,yte
-    else:
-        return xtr,ytr, xte,yte, xte,yte
-
-
-def get_data_torchvision(data_folder = './', dataset = torchvision.datasets.CIFAR10, val_split = 1/5, augment = True):
-    '''
-    Args:
-        dataset (Method from torchvision.datasets): Currently supports MNIST, FMNIST, CIFAR10, CIFAR100
-        val_split (float, optional): What fraction of training data to use for validation
-            If not 0, val data is taken from end of training set. Eg: For val_split=1/5, last 10k images out of 50k for CIFAR are taken as val
-            If 0, train set is complete train set (including val). Test data is returned as val set
-            Defaults to 1/5
-        augment (bool, optional): If True, do transformations
-            Defaults to True
-    
-    Returns:
-        dict: Keys 'train', 'val', 'test'
-            Each is a dataset object which can be fed to a data loader
-            If val_split is 0, test data is returned as 'val'
-
-    '''
-    ## All transforms ##
-    if dataset == torchvision.datasets.MNIST:
-        transform_train = transforms.Compose([transforms.ToTensor()])
-        transform_valtest = transforms.Compose([transforms.ToTensor()])
-        
-    elif dataset == torchvision.datasets.FashionMNIST:
-        if not augment: #no transforms
-            transform_train = transforms.Compose([transforms.ToTensor()])
-            transform_valtest = transforms.Compose([transforms.ToTensor()])
-        else: #fancy transforms, without changing dataset size
-            mus = (0.2855,)
-            sigmas = (0.3528,)
-            transform_train = transforms.Compose([
-                    transforms.RandomCrop(28, padding=4),
-                    transforms.RandomHorizontalFlip(),
-                    transforms.ToTensor(),
-                    transforms.Normalize(mus, sigmas),
-                    # transforms.RandomErasing()
-                    ])
-            transform_valtest = transforms.Compose([
-                    transforms.ToTensor(),
-                    transforms.Normalize(mus, sigmas)
-                    ])
-    
-    elif dataset == torchvision.datasets.CIFAR10 or dataset == torchvision.datasets.CIFAR100:
-        if not augment: #no transforms
-            transform_train = transforms.Compose([transforms.ToTensor()])
-            transform_valtest = transforms.Compose([transforms.ToTensor()])
-        else: #fancy transforms, without changing dataset size
-            mus = (0.4914,0.4821,0.4464) if dataset == torchvision.datasets.CIFAR10 else (0.507,0.4867,0.441)
-            sigmas = (0.2471,0.2436,0.2616) if dataset == torchvision.datasets.CIFAR10 else (0.2674,0.2564,0.276)
-            transform_train = transforms.Compose([
-                    transforms.RandomCrop(32, padding=4),
-                    transforms.RandomHorizontalFlip(),
-                    transforms.ToTensor(),
-                    transforms.Normalize(mus, sigmas),
-                    # transforms.RandomErasing()
-                    ])
-            transform_valtest = transforms.Compose([
-                    transforms.ToTensor(),
-                    transforms.Normalize(mus, sigmas)
-                    ])
-        
-    ## Create actual datasets ##
-    if val_split != 0: #val data exists
-        train_full = dataset(root=data_folder, train=True, download=True, transform=transform_train)
-        val_full = dataset(root=data_folder, train=True, download=True, transform=transform_valtest)
-        split = int((1-val_split)*len(train_full))
-        
-        train = torch.utils.data.Subset(train_full, range(split))
-        val = torch.utils.data.Subset(val_full, range(split, len(val_full)))
-        test = dataset(root=data_folder, train=False, download=True, transform=transform_valtest)
-        return {'train':train, 'val':val, 'test':test}
-    
-    else: #test is val data, actual val data is combined with train
-        train = dataset(root=data_folder, train=True, download=True, transform=transform_train)
-        test = dataset(root=data_folder, train=False, download=True, transform=transform_valtest)
-        return {'train':train, 'val':test, 'test':test} #doesn't matter that we return test as 'test', it won't be used in that sense
-# =============================================================================
-
-
 # =============================================================================
 # Helper methods
 # =============================================================================
