+++ model_search.py
@@ -10,13 +10,7 @@ import sobol_seq
 import pickle
 import itertools
 import time
-
-import os
-
-if os.environ['DNC_DL_FRAMEWORK'] == 'torch':
-    from model.torch_model import run_network, get_numparams, net_kws_defaults, run_kws_defaults
-elif os.environ['DNC_DL_FRAMEWORK'] == 'tf.keras':
-    from model.tf_model import run_network, get_numparams, net_kws_defaults, run_kws_defaults
+from model.model import run_network, get_numparams, net_kws_defaults, run_kws_defaults
 
 
 # =============================================================================
